fix(cm): component edit view settings
