+++ packages/core/admin/admin/src/content-manager/pages/ComponentSettingsView.tsx
@@ -10,7 +10,7 @@ import { reducer, initialState } from '../sharedReducers/crud/reducer';
 import { mergeMetasWithSchema } from '../utils/schemas';
 
 import { selectSchemas } from './App';
-import { EditSettingsView } from './EditSettingsView/EditSettingsView';
+import { SettingsForm } from './EditSettingsView/components/SettingsForm/SettingsForm';
 
 const ComponentSettingsView = () => {
   const [{ isLoading, data: layout }, dispatch] = React.useReducer(reducer, initialState);
@@ -53,7 +53,7 @@ const ComponentSettingsView = () => {
 
   return (
     <CheckPagePermissions permissions={permissions.contentManager?.componentsConfigurations}>
-      <EditSettingsView />
+      <SettingsForm components={layout.components} layout={layout.component} />
     </CheckPagePermissions>
   );
 };
