+++ packages/core/admin/admin/src/content-manager/pages/EditSettingsView/components/SettingsForm/tests/init.test.ts
