+++ packages/core/admin/admin/src/content-manager/pages/EditSettingsView/components/SettingsForm/tests/reducer.test.ts
